% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMCmixture.R
\name{MCMCmixture}
\alias{MCMCmixture}
\title{MCMC mixture}
\usage{
MCMCmixture(m1, m2, s1, s2, alpha, Nsim, density = "Normal")
}
\arguments{
\item{m1}{(numeric data type): mean of 1st univariate gaussian distribution or median of 1st cauchy Distribution depending on value of density parameter}

\item{m2}{(numeric data type) : mean of 2nd univariate gaussian distribution or median of 2nd cauchy Distribution depending on value of density parameter}

\item{s1}{(numeric data type) : variance of 1st univariate gaussian distribution or scale parameter of 1st Cauchy Distribution depending on value of density parameter}

\item{s2}{(numeric data type) : variance of 2nd univariate gaussian distribution or scale parameter of 2nd Cauchy Distribution depending on value of density parameter}

\item{alpha}{(numeric data type) : value of mixing parameter}

\item{Nsim}{number of simulations of Markov Chain having stationery distribution as the target mixture distribution}

\item{density}{(character data type) : specifies the density function as either Normal or Cauchy. Default = "Normal"}
}
\value{
vector of length 1000 of simulated values from the target distribution
}
\description{
MCMCmixture function simulates
random variables from a Univariate Bimodal Gaussian/Cauchy Mixture Distribution. The means/medians and variances/scale parameters of the 2
Gaussian/Cauchy distributions and the mixing parameter that form the target mixture distribution will be taken
as input from the user, along with the iteration number of the Markov Chains to be simulated. This function generates 1000 replications
of the markov chain with stationery distribution as the target distribution, returns the 1000-long vector of generated values at the Nsim-th
iteration of each chain and plots the histogram of the values produced.
}
\examples{
MCMCmixture(1, 10, 2, 3, 0.3, 1000, density = "Normal")
MCMCmixture(-10, 20, 10, 5, 0.6, 2000, density = "Normal")
MCMCmixture(-1, 7, 1, 3, 0.5, 3000, density = "Cauchy")

}
\references{
W.R. Gilks, S. Richardson, and D.J. Spiegelhalter, ed. (1996),
Markov chain Monte Carlo in practice. Chapman and Hall, London.
}
