% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMCmixtureGaussian.R
\name{MCMCmixtureGaussian}
\alias{MCMCmixtureGaussian}
\title{MCMC mixture Gaussian}
\usage{
MCMCmixtureGaussian(mean1, mean2, var1, var2, alpha, Nsim)
}
\arguments{
\item{mean1}{mean of 1st univariate gaussian distribution}

\item{mean2}{mean of 2nd univariate gaussian distribution}

\item{var1}{variance of 1st univariate gaussian distribution}

\item{var2}{variance of 2nd univariate gaussian distribution}

\item{alpha}{mixture proportion of the normal density with smaller mean}

\item{Nsim}{number of values to be simulated from the mixture distribution}
}
\value{
vector of simulated values from the target distribution
}
\description{
MCMCmixtureGaussian function simulate
random variables from a Univariate Bimodal Gaussian Mixture Distribution. The means and variances of the 2
Gaussian distributions and the mixing parameter that form the target mixture distribution will be taken
as input from the user, along with the number of values to be simulated. This function generates 1000 replications
of the markov chain with stationery distribution as the target distribution and plots the histogram of the values produced.
}
\examples{
MCMCmixtureGaussian(1, 10, 2, 3, 0.3, 1000)
MCMCmixtureGaussian(-10, 20, 10, 5, 0.6, 1000)
MCMCmixtureGaussian(-10, 20, 10, 10, 0.5, 1000)


}
