% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMCrateGaussian.R
\name{MCMCrateGaussian}
\alias{MCMCrateGaussian}
\title{MCMC rate Gaussian}
\usage{
MCMCrateGaussian(m1, m2, s1, s2, alpha, Nsim, density = "Normal")
}
\arguments{
\item{m1}{mean of 1st univariate gaussian distribution or median of 1st cauchy Distribution depending on value of density parameter}

\item{m2}{mean of 2nd univariate gaussian distribution or median of 2nd cauchy Distribution depending on value of density parameter}

\item{s1}{variance of 1st univariate gaussian distribution or scale parameter of 1st Cauchy Distribution depending on value of density parameter}

\item{s2}{variance of 2nd univariate gaussian distribution or scale parameter of 2nd Cauchy Distribution depending on value of density parameter}

\item{alpha}{value of mixing parameter}

\item{Nsim}{number of values to be simulated from the mixture distribution}

\item{density}{specifies the density function as either Normal or Cauchy}
}
\value{
vector of supremum of the absolute difference between the empirical
CDF of the 500 generated values of the chain and the actual CDF of the
target distribution at each iteration.
}
\description{
MCMCrateGaussian generate 500 Markov Chain values simultaneously
at each iteration of the Algorithm. As its output, it will return the curve of supremum of the absolute
difference between the empirical CDF of the 500 generated values of the chain and the actual CDF of the
target distribution at each iteration.
}
\examples{
MCMCrateGaussian(m1 = 20, m2 = 0, s1 = 1, s2 = 2, alpha= 0.5, Nsim= 1000)

}
