% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMCmixture.R
\name{chain}
\alias{chain}
\title{chain}
\usage{
chain(t, m1, m2, s1, s2, alpha, Nsim, density = "Normal")
}
\arguments{
\item{t}{starting value of the chain}

\item{m1}{mean of 1st univariate gaussian distribution or median of 1st cauchy Distribution depending on value of density parameter}

\item{m2}{mean of 2nd univariate gaussian distribution or median of 2nd cauchy Distribution depending on value of density parameter}

\item{s1}{variance of 1st univariate gaussian distribution or scale parameter of 1st Cauchy Distribution depending on value of density parameter}

\item{s2}{variance of 2nd univariate gaussian distribution or scale parameter of 2nd Cauchy Distribution depending on value of density parameter}

\item{alpha}{value of mixing parameter}

\item{Nsim}{number of values to be simulated from the mixture distribution}

\item{density}{specifies the density function as either Normal or Cauchy}
}
\value{
vector of simulated values from the target distribution
}
\description{
chain simulates random variables from a Markov Chain having stationery distribution as
a univariate Bimodal Gaussian Mixture Distribution. The means and variances of the 2
Gaussian distributions and the mixing parameter that form the target mixture distribution will be taken
as input from the user, along with the number of values to be simulated. The function returns
the values generated from the chain
}
\examples{
chain(2, 1, 10, 2, 3, 0.3, 1000)
chain(5, 4, 3, 3, 3, 0.5, 1000)
chain(-3, -2, -4, 2, 2, 0.9, 1000, density = "Cauchy")

}
