% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMCmixtureGaussian.R
\name{chain}
\alias{chain}
\title{chain}
\usage{
chain(t, mean1, mean2, var1, var2, alpha, Nsim)
}
\arguments{
\item{t}{starting value of the chain}

\item{mean1}{mean of 1st univariate gaussian distribution}

\item{mean2}{mean of 2nd univariate gaussian distribution}

\item{var1}{variance of 1st univariate gaussian distribution}

\item{var2}{variance of 2nd univariate gaussian distribution}

\item{alpha}{mixture proportion of the normal density with smaller mean}

\item{Nsim}{number of values to be simulated from the mixture distribution}
}
\value{
vector of simulated values from the target distribution
}
\description{
chain simulates random variables from a Markov Chain having stationery distribution as
a univariate Bimodal Gaussian Mixture Distribution. The means and variances of the 2
Gaussian distributions and the mixing parameter that form the target mixture distribution will be taken
as input from the user, along with the number of values to be simulated. The function returns
the values generated from the chain
}
\examples{
chain(2, 1, 10, 2, 3, 0.3, 1000)
chain(5, 4, 3, 3, 3, 0.5, 1000)
chain(-3, -2, -4, 2, 2, 0.9, 1000)
}
