% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMCMixBimodal-package.R
\docType{package}
\name{MCMCMixBimodal-package}
\alias{MCMCMixBimodal}
\alias{MCMCMixBimodal-package}
\title{MCMCMixBimodal: : MCMC for Bimodal Mixture of 2 Univariate Distributions}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Srijato Bhattacharyya \email{srijato@tamu.edu}

}
\keyword{internal}
