% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMCmixtureGaussian.R
\name{MCMCmixtureGaussian}
\alias{MCMCmixtureGaussian}
\title{MCMC for Bimodal Mixtureof 2 Univariate Distributions}
\usage{
MCMCmixtureGaussian(mean1, mean2, var1, var2, nsim)
}
\arguments{
\item{mean1}{mean of 1st univariate gaussian distribution}

\item{mean2}{mean of 2nd univariate gaussian distribution}

\item{var1}{variance of 1st univariate gaussian distribution}

\item{var2}{variance of 2nd univariate gaussian distribution}

\item{nsim}{number of values to be simulated from the mixture distribution}
}
\description{
MCMC for Bimodal Mixtureof 2 Univariate Distributions
}
