% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMCrateGaussian.R
\name{MCMCrateGaussian}
\alias{MCMCrateGaussian}
\title{MCMC rate Gaussian}
\usage{
MCMCrateGaussian(mean1, mean2, var1, var2, alpha, Nsim)
}
\arguments{
\item{mean1}{mean of 1st univariate gaussian distribution}

\item{mean2}{mean of 2nd univariate gaussian distribution}

\item{var1}{variance of 1st univariate gaussian distribution}

\item{var2}{variance of 2nd univariate gaussian distribution}

\item{alpha}{value of mixing parameter}

\item{Nsim}{number of values to be simulated from the mixture distribution}
}
\value{
plot of supremum of the absolute
difference between the empirical CDF of the 500 generated values of the chain and the actual CDF of the
target distribution at each iteration.
}
\description{
MCMCrateGaussian generate 500 Markov Chain values simultaneously
at each iteration of the Algorithm. As its output, it will return the curve of supremum of the absolute
difference between the empirical CDF of the 500 generated values of the chain and the actual CDF of the
target distribution at each iteration.
}
\examples{
MCMCrateGaussian(mean1 = 20, mean2 = 0, var1= 1, var2 = 2, alpha= 0.5, Nsim= 1000)

}
