% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMCmixtureGaussian.R
\name{target_density}
\alias{target_density}
\title{target_density}
\usage{
target_density(x, mean1, mean2, var1, var2, alpha)
}
\arguments{
\item{x}{argument value of the density function}

\item{mean1}{mean of 1st univariate gaussian distribution}

\item{mean2}{mean of 2nd univariate gaussian distribution}

\item{var1}{variance of 1st univariate gaussian distribution}

\item{var2}{variance of 2nd univariate gaussian distribution}

\item{alpha}{value of mixing parameter}
}
\value{
value of the density function at x
}
\description{
target_density gives the value of the density of a univariate Bimodal Gaussian Mixture Distribution. The means and variances of the 2
Gaussian distributions and the mixing parameter are taken as input from the user.
}
