% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMCmixture.R
\name{target_density}
\alias{target_density}
\title{target_density}
\usage{
target_density(x, m1, m2, s1, s2, alpha, density = "Normal")
}
\arguments{
\item{x}{argument value of the density function}

\item{m1}{mean of 1st univariate gaussian distribution or median of 1st cauchy Distribution depending on value of density parameter}

\item{m2}{mean of 2nd univariate gaussian distribution or median of 2nd cauchy Distribution depending on value of density parameter}

\item{s1}{variance of 1st univariate gaussian distribution or scale parameter of 1st Cauchy Distribution depending on value of density parameter}

\item{s2}{variance of 2nd univariate gaussian distribution or scale parameter of 2nd Cauchy Distribution depending on value of density parameter}

\item{alpha}{value of mixing parameter}

\item{density}{specifies the density function as either Normal or Cauchy}
}
\value{
value of the density function at x
}
\description{
target_density gives the value of the density of a univariate Bimodal Gaussian / Bimodal Cauchy Mixture Distribution. The means/medians and variances/scale parameters of the 2
Gaussian/Cauchy distributions and the mixing parameter are taken as input from the user.
}
\examples{
target_density(x = 1, m1 = 5, m2 = 0, s1 = 3, s2 = 1, alpha = 0.5)
target_density(x = 8, m1 = 2, m2 = 7, s1 = 2, s2 = 1, alpha = 0.1, density = "Normal")
target_density(x = 7, m1 = 5, m2 = 1, s1 = 3, s2 = 2, alpha = 0.3, density = "Cauchy")

}
